Extract ControllerState class to its own file
