+++ osu.Framework/Platform/Sdl/Sdl2WindowBackend.cs
@@ -36,7 +36,7 @@ namespace osu.Framework.Platform.Sdl
         private bool mouseInWindow;
         private Point previousPolledPoint = Point.Empty;
 
-        private readonly Dictionary<int, ControllerState> controllers = new Dictionary<int, ControllerState>();
+        private readonly Dictionary<int, Sdl2ControllerState> controllers = new Dictionary<int, Sdl2ControllerState>();
 
         #region Internal Properties
 
@@ -683,7 +683,7 @@ namespace osu.Framework.Platform.Sdl
                     var controller = SDL.SDL_GameControllerOpen(evtCdevice.which);
                     var joystick = SDL.SDL_GameControllerGetJoystick(controller);
                     var instanceID = SDL.SDL_JoystickGetDeviceInstanceID(evtCdevice.which);
-                    controllers[instanceID] = new ControllerState(instanceID, joystick, controller);
+                    controllers[instanceID] = new Sdl2ControllerState(instanceID, joystick, controller);
                     break;
 
                 case SDL.SDL_EventType.SDL_CONTROLLERDEVICEREMOVED:
@@ -730,7 +730,7 @@ namespace osu.Framework.Platform.Sdl
                         break;
 
                     var joystick = SDL.SDL_JoystickOpen(evtJdevice.which);
-                    controllers[instanceID] = new ControllerState(instanceID, joystick, IntPtr.Zero);
+                    controllers[instanceID] = new Sdl2ControllerState(instanceID, joystick, IntPtr.Zero);
                     break;
 
                 case SDL.SDL_EventType.SDL_JOYDEVICEREMOVED:
@@ -1022,71 +1022,6 @@ namespace osu.Framework.Platform.Sdl
 
                 case SDL.SDL_GameControllerAxis.SDL_CONTROLLER_AXIS_TRIGGERRIGHT:
                     return JoystickAxisSource.Axis6;
-            }
-        }
-
-        protected class ControllerState
-        {
-            public readonly int InstanceID;
-            public readonly IntPtr JoystickHandle;
-            public readonly IntPtr ControllerHandle;
-
-            /// <summary>
-            /// Bindings returned from <see cref="SDL.SDL_GameControllerGetBindForButton"/>, indexed by <see cref="SDL.SDL_GameControllerButton"/>.
-            /// Empty if the joystick does not have a corresponding ControllerHandle.
-            /// </summary>
-            public SDL.SDL_GameControllerButtonBind[] ButtonBindings;
-
-            /// <summary>
-            /// Bindings returned from <see cref="SDL.SDL_GameControllerGetBindForAxis"/>, indexed by <see cref="SDL.SDL_GameControllerAxis"/>.
-            /// Empty if the joystick does not have a corresponding ControllerHandle.
-            /// </summary>
-            public SDL.SDL_GameControllerButtonBind[] AxisBindings;
-
-            public JoystickButton[] AxisDirectionButtons;
-
-            public ControllerState(int instanceID, IntPtr joystickHandle, IntPtr controllerHandle)
-            {
-                InstanceID = instanceID;
-                JoystickHandle = joystickHandle;
-                ControllerHandle = controllerHandle;
-                AxisDirectionButtons = new JoystickButton[(int)JoystickAxisSource.AxisCount];
-
-                PopulateBindings();
-            }
-
-            public void PopulateBindings()
-            {
-                if (ControllerHandle == IntPtr.Zero)
-                    return;
-
-                ButtonBindings = Enumerable.Range(0, (int)SDL.SDL_GameControllerButton.SDL_CONTROLLER_BUTTON_MAX)
-                                           .Select(i => SDL.SDL_GameControllerGetBindForButton(ControllerHandle, (SDL.SDL_GameControllerButton)i)).ToArray();
-
-                AxisBindings = Enumerable.Range(0, (int)SDL.SDL_GameControllerAxis.SDL_CONTROLLER_AXIS_MAX)
-                                         .Select(i => SDL.SDL_GameControllerGetBindForAxis(ControllerHandle, (SDL.SDL_GameControllerAxis)i)).ToArray();
-            }
-
-            public SDL.SDL_GameControllerButton GetButtonForIndex(byte index)
-            {
-                for (var i = 0; i < ButtonBindings.Length; i++)
-                {
-                    if (ButtonBindings[i].bindType != SDL.SDL_GameControllerBindType.SDL_CONTROLLER_BINDTYPE_NONE && ButtonBindings[i].value.button == index)
-                        return (SDL.SDL_GameControllerButton)i;
-                }
-
-                return SDL.SDL_GameControllerButton.SDL_CONTROLLER_BUTTON_INVALID;
-            }
-
-            public SDL.SDL_GameControllerAxis GetAxisForIndex(byte index)
-            {
-                for (var i = 0; i < AxisBindings.Length; i++)
-                {
-                    if (AxisBindings[i].bindType != SDL.SDL_GameControllerBindType.SDL_CONTROLLER_BINDTYPE_NONE && AxisBindings[i].value.button == index)
-                        return (SDL.SDL_GameControllerAxis)i;
-                }
-
-                return SDL.SDL_GameControllerAxis.SDL_CONTROLLER_AXIS_INVALID;
             }
         }
